Enable QML debugging
Background:
By accident I ruined my qtcreator environment and had to set up cmake/build
environment from scratch. Since then it was not possible to start QML-Debug
Session: A message box with 'could not connect to the in-process qml debugger'
popped up. Found [1] and checked header QQmlDebuggingEnabler for #if...
With this modifcation debugging works as it used to.

[1] https://forum.qt.io/topic/92019/could-not-connect-to-the-in-process-qml-profiler

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>