MultipleErrorView: Introduce a magic auto-scroll
The solution is simple: When new rows are added we check if scroll position
was at end: If yes: Do scroll otherwise not. So if the user wants to see
upper results, we keep position as it would be annoying to override user.
As soon as user scrolls back to end, auto-scrolling starts again.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>