Get AUX values out of auto scaling when not displayed
Background: Most displays of ActualValueModel do not show aux values or show
them just optionally.
On that cases auto scale is working unexpected: E.g if AUX values are much
bigger than L1/L2/L3 (try ratio to get that) the scaling seems to malfunction.
To not repeat ourselves, we introduce two object of ActualValueModel and
ActualValueModelWithAux.
The solution tries to keep changes limited - things in here are already verbose
and uneasy to understand.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>